Update GUI.py
Added in support for the case when game first starts up and player doesn't choose a hand to throw but presses throw. Previously an error would occur in the background due to < and > comparisons to None.
Now first checks if playerHand is None,  and if so displays text that player needs to choose a hand.